+++ scripts/help.js
@@ -326,7 +326,7 @@ export function checkModules() {
         ui.notifications.error('Chris\'s Premades | You do not have a JB2A module active, this module requires either the free or patreon version', {permanent: true});
     }
     if (game.modules.get('dae')?.active && !game.settings.get('dae', 'showInline')) {
-        ui.notifications.warn('Chris\'s Premades | Dynamic Active Effects setting "Display results of inline rolls" is off, it is recommeded to enable it.', {permanent: true});
+        ui.notifications.warn('Chris\'s Premades | Dynamic Active Effects setting "Display results of inline rolls" is off, it is recommeded to enable it.');
     }
     if (game.modules.get('dfreds-convenient-effects')?.active && (game.settings.get('dfreds-convenient-effects', 'modifyStatusEffects') === 'add')) {
         ui.notifications.warn('Chris\'s Premades | Dfred\'s Convenient Effects setting "Modify Status Effects" is set to "Add", it is recommeded to set to "Replace".', {permanent: true});
@@ -338,7 +338,7 @@ export function checkModules() {
         game.settings.get('midi-qol', 'ConfigSettings').mergeCard,
         game.settings.get('midi-qol', 'ConfigSettings').allowActorUseMacro,
         game.settings.get('midi-qol', 'ConfigSettings').allowUseMacros
-    ]
+    ];
     if (midiSettings.includes(false)) {
         ui.notifications.warn('Chris\'s Premades | Problematic Midi-QOL settings detected, go to CPR\'s settings under help and click "Apply recommended setting fixes".', {permanent: true});
     }
@@ -354,5 +354,8 @@ export function checkModules() {
         if (!game.modules.get(mod.id)?.active) {
             ui.notifications.error('Chris\'s Premades | <b>REQUIRED MODULE</b> "<u>' + mod.title + '</u>" is disabled, please enable.', {permanent: true});
         }
+    }
+    if (Array.from(game.settings.settings).filter(i => i[1].namespace === 'chris-premades').filter(i => !i[1].key.toLowerCase().includes('randomizer')).filter(i => game.settings.get('chris-premades', i[1].key) === false).length === 0) {
+        ui.notifications.error('Chris\'s Premades | You have turned on <b>EVERY CPR setting</b>, please check that you have <u>only enabled what you need</u>.', {permanent: true});
     }
 }