+++ scripts/helperFunctions.js
@@ -1357,19 +1357,22 @@ export let chris = {
         let check = getProperty(game.permissions, permission);
         return check ? check.includes(user.role) : false;
     },
-    'useSpellWhenEmpty': async function _useSpellWhenEmpty(workflow, dialogTitle, dialogMessage, options = {minLevel: 0, dialogOnly: false, consumeSlotOnly: false}) {
-        if (workflow.item?.uses?.value === 0) {
+    'useSpellWhenEmpty': async function _useSpellWhenEmpty(workflow, dialogTitle, dialogMessage, options = {minLevel: 0, dialogOnly: false, consumeSlotOnly: false, skipEmptyCheck: false}) {
+        if (workflow.item?.system?.uses?.value === 0 || options?.skipEmptyCheck) {
             let spellSlots = workflow.actor.system.spells;
             delete spellSlots.spell0;
-            spellSlots = spellSlots.filter(spell => spell.max != 0);
-            if (options?.minLevel) spellSlots = spellSlots.filter(spell => spell.level < options?.minLevel);
+            for (let [key, value] of Object.entries(spellSlots)) {
+                if (value.max === 0 || value.value === 0) delete spellSlots[key];
+                else if (options?.minLevel && value.level < options?.minLevel) delete spellSlots[key]; 
+            }
             let buttons = [];
             for (let [key, value] of Object.entries(spellSlots)) {
                 buttons.push([
-                    key === 'pact' ? `Pact Slot (Level ${value.level})` : value.level + value.level === 1 ? 'st' : value.level === 2 ? 'nd' : value.level === 3 ? 'rd' : 'th' + ' Level Slot' + '(' + value.value + ')',
+                    key === 'pact' ? `Pact Slot (Level ${value.level})` : (value.level + (value.level === 1 ? 'st' : value.level === 2 ? 'nd' : value.level === 3 ? 'rd' : 'th') + ' Level Slot') + ' (' + value.value + ')',
                     key
                 ]);
             }
+            buttons.push(['No', false]);
             let result = await chris.dialog(dialogTitle, buttons, dialogMessage);
             if (!result) return false;
             if (!options?.dialogOnly) {
@@ -1383,7 +1386,8 @@ export let chris = {
                         'useSpellWhenEmpty': result
                     };
                 }
-                await workflow.actor.update('system.spells.' + result, workflow.actor.system.spells[result].value - 1);
+                console.log(result, 'system.spells.' + result, workflow.actor.system.spells[result].value);
+                await workflow.actor.update({['system.spells.' + result + '.value']: workflow.actor.system.spells[result].value - 1});
             }
             return result;
         } else return false;